Rebase to master and generate apps
