+++ apps/cq-gemfire-source-kafka/src/main/java/org/springframework/cloud/stream/app/cq/gemfire/source/kafka/CqGemfireSourceKafkaApplication.java
@@ -1,12 +1,12 @@
-package org.springframework.cloud.stream.app.gemfire.source;
+package org.springframework.cloud.stream.app.cq.gemfire.source.kafka;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class GemfireSourceRabbitApplication {
+public class CqGemfireSourceKafkaApplication {
 
 	public static void main(String[] args) {
-		SpringApplication.run(GemfireSourceRabbitApplication.class, args);
+		SpringApplication.run(CqGemfireSourceKafkaApplication.class, args);
 	}
 }
