+++ apps/cq-gemfire-source-kafka/src/test/java/org/springframework/cloud/stream/app/cq/gemfire/source/kafka/CqGemfireSourceKafkaApplicationTests.java
@@ -1,15 +1,13 @@
-package org.springframework.cloud.stream.app.cq.gemfire.source;
+package org.springframework.cloud.stream.app.cq.gemfire.source.kafka;
 
-import org.junit.Ignore;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.springframework.boot.test.SpringApplicationConfiguration;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 
-@Ignore
 @RunWith(SpringJUnit4ClassRunner.class)
-@SpringApplicationConfiguration(classes = CqGemfireSourceRabbitApplication.class)
-public class CqGemfireSourceRabbitApplicationTests {
+@SpringApplicationConfiguration(classes = CqGemfireSourceKafkaApplication.class)
+public class CqGemfireSourceKafkaApplicationTests {
 
 	@Test
 	public void contextLoads() {
