+++ apps/cq-gemfire-source-rabbit/src/main/java/org/springframework/cloud/stream/app/cq/gemfire/source/rabbit/CqGemfireSourceRabbitApplication.java
@@ -1,4 +1,4 @@
-package org.springframework.cloud.stream.app.cq.gemfire.source;
+package org.springframework.cloud.stream.app.cq.gemfire.source.rabbit;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
