+++ apps/cq-gemfire-source-rabbit/src/test/java/org/springframework/cloud/stream/app/cq/gemfire/source/rabbit/CqGemfireSourceRabbitApplicationTests.java
@@ -1,15 +1,13 @@
-package org.springframework.cloud.stream.app.gemfire.source;
+package org.springframework.cloud.stream.app.cq.gemfire.source.rabbit;
 
-import org.junit.Ignore;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.springframework.boot.test.SpringApplicationConfiguration;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 
-@Ignore
 @RunWith(SpringJUnit4ClassRunner.class)
-@SpringApplicationConfiguration(classes = GemfireSourceKafkaApplication.class)
-public class GemfireSourceKafkaApplicationTests {
+@SpringApplicationConfiguration(classes = CqGemfireSourceRabbitApplication.class)
+public class CqGemfireSourceRabbitApplicationTests {
 
 	@Test
 	public void contextLoads() {
