+++ apps/gemfire-sink-kafka/pom.xml
@@ -9,10 +9,10 @@
     <relativePath></relativePath>
   </parent>
   <groupId>org.springframework.cloud.stream.app</groupId>
-  <artifactId>cq-gemfire-source-kafka</artifactId>
+  <artifactId>gemfire-sink-kafka</artifactId>
   <version>1.0.0.BUILD-SNAPSHOT</version>
-  <name>cq-gemfire-source-kafka</name>
-  <description>Spring Cloud Stream Cq Gemfire Source Kafka Binder Application</description>
+  <name>gemfire-sink-kafka</name>
+  <description>Spring Cloud Stream Gemfire Sink Kafka Binder Application</description>
   <properties>
     <skipTests>true</skipTests>
     <java.version>1.7</java.version>
@@ -32,7 +32,7 @@
   <dependencies>
     <dependency>
       <groupId>org.springframework.cloud.stream.app</groupId>
-      <artifactId>spring-cloud-starter-stream-source-gemfire-cq</artifactId>
+      <artifactId>spring-cloud-starter-stream-sink-gemfire</artifactId>
     </dependency>
     <dependency>
       <groupId>org.springframework.cloud</groupId>
@@ -93,7 +93,7 @@
                   <exec>
                     <arg>java</arg>
                     <arg>-jar</arg>
-                    <arg>/maven/cq-gemfire-source-kafka.jar</arg>
+                    <arg>/maven/gemfire-sink-kafka.jar</arg>
                   </exec>
                 </entryPoint>
                 <assembly>
