+++ apps/gemfire-sink-kafka/src/main/docker/assembly.xml
@@ -1,0 +1,15 @@
+<assembly
+        xmlns="http://maven.apache.org/plugins/maven-assembly-plugin/assembly/1.1.2"
+        xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+        xsi:schemaLocation="http://maven.apache.org/plugins/maven-assembly-plugin/assembly/1.1.2 http://maven.apache.org/xsd/assembly-1.1.2.xsd">
+    <id>gemfire-sink-kafka</id>
+    <dependencySets>
+        <dependencySet>
+            <includes>
+                <include>org.springframework.cloud.stream.app:gemfire-sink-kafka</include>
+            </includes>
+            <outputDirectory>.</outputDirectory>
+            <outputFileNameMapping>gemfire-sink-kafka.jar</outputFileNameMapping>
+        </dependencySet>
+    </dependencySets>
+</assembly>
