+++ apps/gemfire-sink-kafka/src/main/java/org/springframework/cloud/stream/app/gemfire/sink/kafka/GemfireSinkKafkaApplication.java
@@ -1,12 +1,12 @@
-package org.springframework.cloud.stream.app.gemfire.source;
+package org.springframework.cloud.stream.app.gemfire.sink.kafka;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class GemfireSourceKafkaApplication {
+public class GemfireSinkKafkaApplication {
 
 	public static void main(String[] args) {
-		SpringApplication.run(GemfireSourceKafkaApplication.class, args);
+		SpringApplication.run(GemfireSinkKafkaApplication.class, args);
 	}
 }
