+++ apps/gemfire-sink-kafka/src/test/java/org/springframework/cloud/stream/app/gemfire/sink/kafka/GemfireSinkKafkaApplicationTests.java
@@ -1,0 +1,16 @@
+package org.springframework.cloud.stream.app.gemfire.sink.kafka;
+
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.springframework.boot.test.SpringApplicationConfiguration;
+import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
+
+@RunWith(SpringJUnit4ClassRunner.class)
+@SpringApplicationConfiguration(classes = GemfireSinkKafkaApplication.class)
+public class GemfireSinkKafkaApplicationTests {
+
+	@Test
+	public void contextLoads() {
+	}
+
+}
