+++ apps/gemfire-sink-rabbit/src/main/java/org/springframework/cloud/stream/app/gemfire/sink/rabbit/GemfireSinkRabbitApplication.java
@@ -1,0 +1,12 @@
+package org.springframework.cloud.stream.app.gemfire.sink.rabbit;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class GemfireSinkRabbitApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(GemfireSinkRabbitApplication.class, args);
+	}
+}
