+++ apps/gemfire-sink-rabbit/src/test/java/org/springframework/cloud/stream/app/gemfire/sink/rabbit/GemfireSinkRabbitApplicationTests.java
@@ -1,15 +1,13 @@
-package org.springframework.cloud.stream.app.gemfire.source;
+package org.springframework.cloud.stream.app.gemfire.sink.rabbit;
 
-import org.junit.Ignore;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.springframework.boot.test.SpringApplicationConfiguration;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 
-@Ignore
 @RunWith(SpringJUnit4ClassRunner.class)
-@SpringApplicationConfiguration(classes = GemfireSourceRabbitApplication.class)
-public class GemfireSourceRabbitApplicationTests {
+@SpringApplicationConfiguration(classes = GemfireSinkRabbitApplication.class)
+public class GemfireSinkRabbitApplicationTests {
 
 	@Test
 	public void contextLoads() {
