+++ apps/gemfire-source-kafka/pom.xml
@@ -14,6 +14,7 @@
   <name>gemfire-source-kafka</name>
   <description>Spring Cloud Stream Gemfire Source Kafka Binder Application</description>
   <properties>
+    <skipTests>true</skipTests>
     <java.version>1.7</java.version>
     <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
   </properties>
@@ -101,6 +102,13 @@
               </build>
             </image>
           </images>
+        </configuration>
+      </plugin>
+      <plugin>
+        <artifactId>maven-surefire-plugin</artifactId>
+        <version>2.19.1</version>
+        <configuration>
+          <skipTests>${skipTests}</skipTests>
         </configuration>
       </plugin>
     </plugins>
