+++ apps/gemfire-source-kafka/src/main/java/org/springframework/cloud/stream/app/gemfire/source/kafka/GemfireSourceKafkaApplication.java
@@ -1,12 +1,12 @@
-package org.springframework.cloud.stream.app.cq.gemfire.source;
+package org.springframework.cloud.stream.app.gemfire.source.kafka;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class CqGemfireSourceKafkaApplication {
+public class GemfireSourceKafkaApplication {
 
 	public static void main(String[] args) {
-		SpringApplication.run(CqGemfireSourceKafkaApplication.class, args);
+		SpringApplication.run(GemfireSourceKafkaApplication.class, args);
 	}
 }
