+++ apps/gemfire-source-kafka/src/test/java/org/springframework/cloud/stream/app/gemfire/source/kafka/GemfireSourceKafkaApplicationTests.java
@@ -1,0 +1,16 @@
+package org.springframework.cloud.stream.app.gemfire.source.kafka;
+
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.springframework.boot.test.SpringApplicationConfiguration;
+import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
+
+@RunWith(SpringJUnit4ClassRunner.class)
+@SpringApplicationConfiguration(classes = GemfireSourceKafkaApplication.class)
+public class GemfireSourceKafkaApplicationTests {
+
+	@Test
+	public void contextLoads() {
+	}
+
+}
