+++ apps/gemfire-source-rabbit/src/main/java/org/springframework/cloud/stream/app/gemfire/source/rabbit/GemfireSourceRabbitApplication.java
@@ -1,0 +1,12 @@
+package org.springframework.cloud.stream.app.gemfire.source.rabbit;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class GemfireSourceRabbitApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(GemfireSourceRabbitApplication.class, args);
+	}
+}
