+++ apps/gemfire-source-rabbit/src/test/java/org/springframework/cloud/stream/app/gemfire/source/rabbit/GemfireSourceRabbitApplicationTests.java
@@ -1,15 +1,13 @@
-package org.springframework.cloud.stream.app.cq.gemfire.source;
+package org.springframework.cloud.stream.app.gemfire.source.rabbit;
 
-import org.junit.Ignore;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.springframework.boot.test.SpringApplicationConfiguration;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 
-@Ignore
 @RunWith(SpringJUnit4ClassRunner.class)
-@SpringApplicationConfiguration(classes = CqGemfireSourceKafkaApplication.class)
-public class CqGemfireSourceKafkaApplicationTests {
+@SpringApplicationConfiguration(classes = GemfireSourceRabbitApplication.class)
+public class GemfireSourceRabbitApplicationTests {
 
 	@Test
 	public void contextLoads() {
